refactor: address architecture review findings
